move auxiliary basis set files
