+++ include/simde/orbital_spaces/orbital_spaces.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "simde/orbital_spaces/auxiliary_basis_set.hpp"
 #include "simde/orbital_spaces/complimentary_auxiliary_basis_set.hpp"
 #include "simde/orbital_spaces/density_fit_basis.hpp"
 #include "simde/orbital_spaces/localized_orbitals.hpp"
