+++ include/simde/simde.hpp
@@ -1,7 +1,6 @@
 #pragma once
 #include "simde/atom_to_ao.hpp"
 #include "simde/atomic_charges.hpp"
-#include "simde/auxiliary_basis_set.hpp"
 #include "simde/connectivity.hpp"
 #include "simde/density/density.hpp"
 #include "simde/energy/energy.hpp"
