+++ tests/simde/orbital_spaces/auxiliary_basis_set.cpp
@@ -1,5 +1,5 @@
-#include "simde/auxiliary_basis_set.hpp"
-#include "test_property_type.hpp"
+#include "../test_property_type.hpp"
+#include "simde/orbital_spaces/auxiliary_basis_set.hpp"
 
 TEST_CASE("AuxiliaryBasisSet") {
     test_property_type<simde::AuxiliaryBasisSet>({"AtomicBasisSet"},
